updated to include GUI
Updated the entire program to work around a GUI system so the algorithms could be visualized and changed, with more ability to change the maze generated or to create your own maze. The ability to see the heatmap of the agents visits on the grid itself and to change the parameters when specified.